+++ tests/BlazingSQLTest/Utils/__init__.py
@@ -97,15 +97,15 @@ def try_to_get_dask_client(n_workers, n_gpus, iface):
             raise ValueError("n gpus must be at least 1")
         try:
             if daskConnection == "local":
-                # cluster = LocalBlazingSQLCluster(n_gpus, n_workers=n_workers)
-                cluster = LocalBlazingSQLCluster(n_gpus,  n_workers=n_workers,
-                    interface=iface,
-                    protocol="ucx",
-                    enable_tcp_over_ucx=True,
-                    enable_infiniband=False,
-                    enable_nvlink=False,
-                    # asynchronous=True,
-                )
+                cluster = LocalBlazingSQLCluster(n_gpus, n_workers=n_workers)
+                # cluster = LocalBlazingSQLCluster(n_gpus,  n_workers=n_workers,
+                #   interface=iface,
+                #    protocol="ucx",
+                #    enable_tcp_over_ucx=True,
+                #    enable_infiniband=False,
+                #    enable_nvlink=False,
+                #    # asynchronous=True,
+                #)
                 return Client(cluster)
             else:
                 if daskConnection.endswith('.json'): #assuming a scheduler file