+++ CHANGELOG.md
@@ -27,7 +27,8 @@
 - #1078 Bump junit from 4.12 to 4.13.1 in /algebra
 - #1144 update with changes from main
 - #1156 Added scheduler file support for e2e testing framework
-- #1158 Deprecated bc.partition
+- #1158 Deprecated bc.partition 
+- #1154 Recompute the avg_bytes_per_row value
 - #1155 Removing comms subproject and cleaning some related code
 
 
@@ -42,7 +43,9 @@
 - #1118 Fixed issue with config_options and adding local_files to valid params
 - #1133 Fixed adressing issue in float columns when parsing parquet metadata
 - #1163 added empty line to trigger build
-- #1168 Fix SSL errors for conda
+- #1168 Fix SSL errors for conda 
+- #1165 E2e tests, distributed mode, again tcp
+
 
 # BlazingSQL 0.16.0 (October 22, 2020)